feat: the Host Tab Plot is on its own container now
