+++ src/components/Explorer/Base/Result/SerratusApiCalls.tsx
@@ -156,8 +156,6 @@ export const fetchPagedHostMatches = async (
         payload
     )
 
-    console.log('response.data', response.data)
-
     return response.data as ResultPagination
 }
