+++ src/components/Host/Host.tsx
@@ -5,31 +5,18 @@ import { RunData } from './types'
 // import { helpIcon } from 'common'
 import { fetchPagedHostMatches } from 'components/Explorer/Base/Result/SerratusApiCalls'
 import { LoadIcon } from 'common/LoadIcon'
-import { axisBottom, axisLeft, scaleBand, scaleLinear, select } from 'd3'
 import { HostTaxIdResult } from './HostTaxIdResult'
+import { COLOR_SCALE, HostTaxPlot } from './HostTaxPlot'
 
 type Props = {
     runIds?: string[]
     isEmbedded?: boolean
 }
 
-// Tableau10 @ https://github.com/d3/d3-scale-chromatic
-const COLOR_SCALE = [
-    '#4e79a7',
-    '#f28e2c',
-    '#e15759',
-    '#76b7b2',
-    '#59a14f',
-    '#edc949',
-    '#af7aa1',
-    '#ff9da7',
-    '#9c755f',
-    '#bab0ab',
-]
-
 export const Host = ({ runIds, isEmbedded = false }: Props) => {
     const [isFetching, setIsFetching] = React.useState<boolean>(true)
     const [runData, setRunData] = React.useState<RunData[]>([])
+    const [taxPlotRunData, setTaxPlotRunData] = React.useState([]);
     const hostBarPlotLegendRef = React.useRef()
 
     React.useEffect(() => {
@@ -95,57 +82,10 @@ export const Host = ({ runIds, isEmbedded = false }: Props) => {
                 return v
             })
 
-            const margin = { top: 32, right: 8, bottom: 64, left: 256 }
-            const width = 800 - margin.left - margin.right
-            const height = 600 - margin.top - margin.bottom
+            setTaxPlotRunData(data)
 
-            const svg = select('#hostBarPlot')
-                .append('svg')
-                .attr('width', width + margin.left + margin.right)
-                .attr('height', height + margin.top + margin.bottom)
-                .append('g')
-                .attr('transform', `translate(${margin.left}, ${margin.top})`)
-
-            svg.append('defs')
-                .append('pattern')
-                .attr('id', `bar-pattern`)
-                .attr('height', 4)
-                .attr('width', 4)
-                .attr('patternUnits', 'userSpaceOnUse')
-                .append('rect')
-                .attr('height', 2)
-                .attr('width', 2)
-                .attr('fill', COLOR_SCALE[9])
-
-            const x = scaleLinear().domain([0, data[0][1]['srarun.run'].length]).range([0, width])
-
-            svg.append('g')
-                .attr('transform', `translate(0, ${height})`)
-                .call(axisBottom(x))
-                .selectAll('text')
-                .style('font-size', '14px')
-                .style('text-anchor', 'middle')
-
-            const y = scaleBand()
-                .range([0, height])
-                .domain(data.map((v: any) => v[1]['srarun.scientific_name']))
-                .padding(0.1)
-
-            svg.append('g').call(axisLeft(y)).style('font-size', '14px')
-
-            svg.selectAll()
-                .data(data)
-                .join('rect')
-                .attr('x', x(0))
-                .attr('y', (d: any) => (y as any)(d[1]['srarun.scientific_name']))
-                .attr('width', (d: any) => (x as any)(d[1]['srarun.run'].length))
-                .attr('height', y.bandwidth())
-                .attr('fill', (d: any) => d[1].color)
-
-            if (hostBarPlotLegendRef.current) {
-                while ((hostBarPlotLegendRef.current as any).childNodes.length)
-                    (hostBarPlotLegendRef.current as any).childNodes[0].remove()
-
+            // X vvvv
+            if (hostBarPlotLegendRef.current)
                 ReactDOM.render(
                     <div style={{ display: 'flex', flexWrap: 'wrap', gap: '16px' }}>
                         {orderList.map((v: any, i: number) => {
@@ -174,9 +114,8 @@ export const Host = ({ runIds, isEmbedded = false }: Props) => {
                     </div>,
                     hostBarPlotLegendRef.current
                 )
-            }
         }
-    }, [runDataGroupEntries])
+    }, [runData])
 
     return (
         <div className='mx-14 my-2'>
@@ -198,7 +137,10 @@ export const Host = ({ runIds, isEmbedded = false }: Props) => {
                         </div>
                     </div>
                     <div className='my-4'>
-                        <div className='w-full text-center' id='hostBarPlot'></div>
+                        {/* <div className='w-full text-center' id='hostBarPlot'></div> */}
+                        <div className='w-full text-center'>
+                            <HostTaxPlot taxPlotRunData={taxPlotRunData} />
+                        </div>
                         <div
                             className='w-full text-center'
                             ref={hostBarPlotLegendRef as LegacyRef<any>}></div>
